NF: move UserDataMigrationPreferences to MigrateUserData package
